:lipstick: :bento: :arrow_up: :heavy_plus_sign: Upgrading Image Gallery library and styles. #29 #17
Upgrading the image gallery library, which required to update the styles and make sure it is not overridden by current css files. 
Adding library to check the device the user is using to make device-dependent changes. Changing the default image in case of image loading error to a more efficient format. 
Removing subtitles.